refactor, realize why having a packed-buffer with the loose db is valuable (#263)
Because that way, the error type can be simpler as it contains errors
related to the packed buffer being initialized.

Now it's strangely on the side, but maybe one can also re-arrange errors
entirely.

Seems harder than putting the packed buffer where it belongs.
